#49 Refactor: 소유 여부 확인 함수 포함 클래스 위치 이전
